feat: Add NewToolsetGroupFromTools constructor
Add a new constructor that creates a ToolsetGroup from a list of ServerTools.
Tools are automatically categorized as read or write based on their
ReadOnlyHint annotation, and grouped into toolsets based on the 'toolset'
field in their Meta.

This enables tools to be fully self-describing - registration can now
derive toolset membership and read/write classification directly from
the tool definition.